Created game statistics module for more organized code
